edit insert function and delete function CMND to upload  image and thumb

+++ nukeviet-develop/modules/cmnd/admin/delete_CMND.php
@@ -4,15 +4,37 @@
 if( ! defined( 'NV_IS_FILE_ADMIN' ) )
 	die( 'Stop!!!' );
 
+//get cmnd_code from delete_CMND.tpl and delete this row 
+$cmnd_code = $nv_Request->get_title('CMND_Code','get');
+
+$sqldel_cmnd_code = "SELECT CMND_code , name , birthday , sex , image, thumb, hometown , origin , place , ethnic , religious , date_of_issue , where_licensing , characteristics 
+FROM ".$db_config['prefix'] . "_" . NV_LANG_DATA . "_" . $module_data." WHERE CMND_Code='".$cmnd_code."'";
+
+$_querydel_cmnd_code = $db->query($sqldel_cmnd_code);
+$row = $_querydel_cmnd_code->fetch();
+
+
+if( $row['image'] != '' && file_exists( NV_UPLOADS_REAL_DIR . "/" . $row['image'] ) )
+		{
+			//xoa anh goc
+			if( @unlink( NV_UPLOADS_REAL_DIR . "/" . $row['image'] ) )
+			{
+				//xoa anh thum
+				if( $row['thumb'] != '' && file_exists( NV_UPLOADS_REAL_DIR . "/" . $row['thumb'] ) )
+				{
+					@unlink( NV_UPLOADS_REAL_DIR . "/" . $row['thumb'] );
+				}
+				
+				//xoa du lieu trong DB		
+				$sql_delcmnd = "delete from ".$db_config['prefix'] . "_" . NV_LANG_DATA . "_" . $module_data." where CMND_Code = '".$cmnd_code."'";
+				$db->query($sql_delcmnd);
+				}
+			}
+	
 
 //if (!empty($delcmnd_code))
 //{
-	//get cmnd_code from delete_CMND.tpl and delete this row 
-	$cmnd_code = $nv_Request->get_title('CMND_Code','get');
-	$sql_delcmnd = "delete from ".$db_config['prefix'] . "_" . NV_LANG_DATA . "_" . $module_data." where CMND_Code = '".$cmnd_code."'";
-	//var_dump($sql_delcmnd);
-	//die();
-	$db->query($sql_delcmnd);
+	
 	//$db->exec();
 //}
 
@@ -29,12 +51,31 @@ $xtpl->assign( 'OP', $op );
 
 
 //Display cmnd list again
-$sql_cmnd = "SELECT CMND_code , name , birthday , sex , hometown , origin , place , ethnic , religious , date_of_issue , where_licensing , characteristics FROM ".$db_config['prefix'] . "_" . NV_LANG_DATA . "_" . $module_data;
+$sql_cmnd = "SELECT CMND_code , name , birthday , sex , image, thumb, hometown , origin , place , ethnic , religious , date_of_issue , where_licensing , characteristics FROM ".$db_config['prefix'] . "_" . NV_LANG_DATA . "_" . $module_data;
 $_query = $db->query($sql_cmnd); 
+
+@require_once ( NV_ROOTDIR . "/includes/class/image.class.php" );
 
 while($row = $_query->fetch())
 {	
-	$xtpl->assign('DATA', $row);
+		$xtpl->assign('DATA', array(
+		"cmnd_code" => $row['cmnd_code'],
+		"name" => $row['name'],
+		"birthday" => $row['birthday'],
+		"sex" => $row['sex'],
+		"image" =>  NV_BASE_SITEURL . NV_UPLOADS_DIR . "/" . $row['image'],
+	    "thumb" =>  NV_BASE_SITEURL . NV_UPLOADS_DIR . "/" . $row['thumb'],
+		"hometown" => $row['hometown'],
+		"origin" => $row['origin'],
+		"place" => $row['place'],
+		"ethnic" => $row['ethnic'],
+		"religious" => $row['religious'],
+		"date_of_issue" => $row['date_of_issue'],
+		"where_licensing" => $row['where_licensing'],
+		"characteristics" => $row['characteristics']
+		));
+	
+	
 	$xtpl->parse('main.loop');
 }
